requirements.txt
asgiref==3.5.2
Django==4.1.1
setuptools==57.0.0
sqlparse==0.4.2
wheel==0.36.2
psycopg2==2.9.3
djangorestframework-simplejwt
djangorestframework